Adding memory.md game to hacks folder
Includes a basic memory game with a 4x4 grid of emojis that, when matched, increase score. The game includes a score and an attempt counter to keep track of different stats. Goes over the basics of how to create a random game as well as including how to make a canvas in html. Students making this game should already know basic js and html, estimated to be a intermediate to advanced hack.